Create Clickable Columns, Create New Course Screen
Course info is now displayed in clickable columns, and have a border
surrounding them to set them apart. Clicking on a course info column
will display a new screen for information related to a specific course.
There's also a back arrow at the top of the screen to return to the homepage.