Alter App Bar Title Size and Screen Fitting
Made the app bar title's a little larger and added a couple of max size
and max height fills. Also, a current bug is that since the scroll
state is remembered across compositions, scrolling down and then selecting
a course to look at makes the course screen stay scrolled down, but there's
currently no way to scroll back up for some reason, causing the user unable
to return to the previous screen. Something to note about this issue is that
both the top app bar title and top row displaying the course name both stay
scrolled up, but the icon strings at the bottom of the screen stay unaffected.