Add Crowd Sourcing Features
The CourseInfo screen now displays 4 clickable buttons at the bottom of the
which allow the user to report a crowded, empty, rainy, or
windy frisbee golf course. These buttons send an HTTP post request to update
the relevant values which are stored in the Diskit server. These values are
also displayed every time the user switches to the CourseInfo screen, but
ideally will be updated every ~30 seconds without needing to move between
screens.